Implemented classes for working with resumes. I've made adjustments to the rest of the classes and tables.

+++ src/main/java/kg/attractor/job_search/service/impl/UserServiceImpl.java
@@ -34,7 +34,7 @@ public class UserServiceImpl implements UserService {
     }
 
     @Override
-    public UserDto getUserById(Long id) {
+    public UserDto getUserById(Integer id) {
         User user = userDao.getUserById(id)
                 .orElseThrow(UserNotFoundException::new);
         return convertToDto(user);
